initial refactoring of cluster-director-mcp into cluster-director-slurm
