+++ cluster-director-slurm/pkg/tools/tools.go
@@ -1,0 +1,26 @@
+// Copyright 2025 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package tools
+
+import (
+	"cluster-director-mcp/cluster-director-slurm/pkg/config"
+	"cluster-director-mcp/cluster-director-slurm/pkg/tools/cluster"
+
+	"github.com/modelcontextprotocol/go-sdk/mcp"
+)
+
+func Install(s *mcp.Server, c *config.Config) {
+	cluster.Install(s, c)
+}
